+++ src/controllers/health.rs
@@ -1,0 +1,13 @@
+use crate::models::health::HealthResponse;
+use crate::state::AppState;
+use axum::routing::get;
+use axum::{Json, Router};
+
+pub fn create_router() -> Router<AppState> {
+    Router::new()
+        .route("/health", get(health_check))
+}
+
+async fn health_check() -> Json<HealthResponse> {
+    Json(HealthResponse { status: "OK".into() })
+}
