+++ src/models.rs
@@ -1,0 +1,2 @@
+pub mod health;
+pub mod services;
