+++ src/models/health.rs
@@ -1,0 +1,6 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct HealthResponse {
+    pub status: String,
+}
