+++ src/models/services.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct ServiceResponse {
+    pub service: String,
+    pub status: String,
+}
